+++ src/scalar/scalarlogiset.jl
@@ -1,152 +1,4 @@
 using ProgressMeter
-using SoleData: AbstractMultiDataset
-import SoleData: ninstances, nvariables, nmodalities, eachmodality, displaystructure
-import SoleData: instances, concatdatasets
-
-"""
-    islogiseed(dataset)::Bool
-
-A logiseed is a dataset that can be converted to a logiset (e.g., via [`scalarlogiset`](@ref)).
-If the `dataset` is multimodal, the following methods should be defined:
-
-```julia
-    islogiseed(::typeof(dataset)) = true
-    initlogiset(dataset, features; kwargs...)
-    ninstances(dataset)
-    nvariables(dataset)
-    frame(dataset, i_instance::Integer)
-    featvalue(feature::VarFeature, dataset, i_instance::Integer, w::AbstractWorld)
-    varnames(dataset)::Union{Nothing,Vector{<:$(VariableId)}}
-    vareltype(dataset, i_variable::$(VariableId))
-```
-
-If `dataset` is multimodal, the following methods should be defined,
-while its modalities (iterated via `eachmodality`) should provide the methods above:
-
-```julia
-    ismultilogiseed(dataset)
-    nmodalities(dataset)
-    eachmodality(dataset)
-```
-"""
-function islogiseed(dataset)
-    false
-    # return error("Please, provide method islogiseed(dataset::$(typeof(dataset))).")
-end
-function initlogiset#(dataset, features)
-    # return error("Please, provide method initlogiset(dataset::$(typeof(dataset)), features::$(typeof(features))).")
-end
-function ninstances#(dataset)
-    # return error("Please, provide method ninstances(dataset::$(typeof(dataset))).")
-end
-function nvariables#(dataset)
-    # return error("Please, provide method nvariables(dataset::$(typeof(dataset))).")
-end
-function frame#(dataset, i_instance::Integer)
-    # return error("Please, provide method frame(dataset::$(typeof(dataset)), i_instance::Integer).")
-end
-function featvalue#(feature::AbstractFeature, dataset, i_instance::Integer, w)
-    # return error("Please, provide method featvalue(feature::$(typeof(feature)), dataset::$(typeof(dataset)), i_instance::Integer, w::$(typeof(w))).")
-end
-function vareltype#(dataset, i_variable::VariableId)
-    # return error("Please, provide method vareltype(dataset::$(typeof(dataset)), i_variable::VariableId).")
-end
-
-function varnames#(dataset)
-    # return error("Please, provide method featvalue(...).")
-end
-function allworlds(dataset, i_instance::Integer)
-    allworlds(frame(dataset, i_instance))
-end
-
-# Multimodal dataset interface
-function ismultilogiseed(dataset)
-    false
-end
-function nmodalities#(dataset)
-    # return error("Please, provide method nmodalities(dataset::$(typeof(dataset))).")
-end
-function eachmodality#(dataset)
-    # return error("Please, provide method eachmodality(dataset::$(typeof(dataset))).")
-end
-
-function modality(dataset, i_modality)
-    eachmodality(dataset)[i_modality]
-end
-
-function ismultilogiseed(dataset::MultiLogiset)
-    true
-end
-function ismultilogiseed(dataset::AbstractMultiDataset)
-    true
-end
-
-function ismultilogiseed(dataset::Union{AbstractVector,Tuple})
-    length(dataset) > 0 && all(islogiseed, dataset) # && allequal(ninstances, eachmodality(dataset))
-end
-function nmodalities(dataset::Union{AbstractVector,Tuple})
-    @assert ismultilogiseed(dataset) "$(typeof(dataset))"
-    length(dataset)
-end
-function eachmodality(dataset::Union{AbstractVector,Tuple})
-    # @assert ismultilogiseed(dataset) "$(typeof(dataset))"
-    dataset
-end
-function ninstances(dataset::Union{AbstractVector,Tuple})
-    @assert ismultilogiseed(dataset) "$(typeof(dataset))"
-    ninstances(first(dataset))
-end
-
-function instances(
-    dataset::Union{AbstractVector,Tuple},
-    inds::AbstractVector,
-    return_view::Union{Val{true},Val{false}} = Val(false);
-    kwargs...
-)
-    @assert ismultilogiseed(dataset) "$(typeof(dataset))"
-    map(modality->instances(modality, inds, return_view; kwargs...), eachmodality(dataset))
-end
-
-function concatdatasets(datasets::Union{AbstractVector,Tuple}...)
-    @assert all(ismultilogiseed.(datasets)) "$(typeof.(datasets))"
-    @assert allequal(nmodalities.(datasets)) "Cannot concatenate multilogiseed's of type ($(typeof.(datasets))) with mismatching " *
-        "number of modalities: $(nmodalities.(datasets))"
-    out = [concatdatasets([modality(dataset, i_mod) for dataset in datasets]...) for i_mod in 1:nmodalities(first(datasets))]
-    if eltype(datasets) <: Tuple
-        out = Tuple(out)
-    end
-    out
-end
-
-function displaystructure(dataset; indent_str = "", include_ninstances = true, kwargs...)
-    if ismultilogiseed(dataset)
-        pieces = []
-        push!(pieces, "multilogiseed with $(nmodalities(dataset)) modalities ($(humansize(dataset)))")
-        # push!(pieces, indent_str * "├ # modalities:\t$(nmodalities(dataset))")
-        if include_ninstances
-            push!(pieces, indent_str * "├ # instances:\t$(ninstances(dataset))")
-        end
-        # push!(pieces, indent_str * "├ modalitytype:\t$(modalitytype(dataset))")
-        for (i_modality, mod) in enumerate(eachmodality(dataset))
-            out = ""
-            if i_modality == nmodalities(dataset)
-                out *= "$(indent_str)└"
-            else
-                out *= "$(indent_str)├"
-            end
-            out *= "{$i_modality} "
-            # \t\t\t$(humansize(mod))\t(worldtype: $(worldtype(mod)))"
-            out *= displaystructure(mod; indent_str = indent_str * (i_modality == nmodalities(dataset) ? "  " : "│ "), include_ninstances = false, kwargs...)
-            push!(pieces, out)
-        end
-        return join(pieces, "\n")
-    elseif islogiseed(dataset)
-        return "logiseed ($(humansize(dataset)))\n$(dataset)" |> x->"$(replace(x, "\n"=>"$(indent_str)\n"))\n"
-    else
-        return "?? dataset of type $(typeof(dataset)) ($(humansize(dataset))) ??\n$(dataset)\n" |> x->"$(replace(x, "\n"=>"$(indent_str)\n"))\n"
-    end
-end
-
 
 """
     scalarlogiset(dataset, features; kwargs...)
